refactor(internal): extract the error-log parser into its own module
tools_utility.py had grown past AGENTS.md's ~1000-line guidance, and the
error-log parsing was the part that shares nothing with the tool
plumbing but its input string: the line regex and constants,
_ExtractedLines, _extract_log_entries, _parse_error_log_structured and
their helpers now live in error_log_parsing.py. Imports and test patch
targets follow the move.

The transplant is verbatim apart from one line: of the 349 lines moved,
348 are unchanged, and the comment above the line regex is rewritten
because it pointed at _LOG_LEVEL_RE "above" — that constant stays in
tools_utility, so the cross-reference had to be re-anchored.

The raw response building moves into _build_raw_error_log alongside the
_build_structured_error_log it mirrors, which also brings _get_error_log
back under ruff's C901 ceiling.

_LOG_LEVEL_RE stays behind — it belongs to the raw path's level filter,
not to the parser.
