+++ tests/src/e2e/tools/test_logbook.py
@@ -8,7 +8,7 @@ import logging
 import pytest
 from fastmcp.exceptions import ToolError
 
-from ha_mcp.tools.tools_utility import _DEFAULT_TOP_N
+from ha_mcp.tools.error_log_parsing import _DEFAULT_TOP_N
 
 from ..utilities.assertions import MCPAssertions, safe_call_tool
